1. [ReviewListAdapter] 닉네임 설정, 본인이 쓴 리뷰 삭제 기능 구현, 주문별 리뷰 평점 표시, 리뷰 총점 표시, 리뷰 점수별 프로그레스바 적용

+++ app/src/main/java/com/example/orderingproject/Dto/request/ReviewDto.java
@@ -13,5 +13,5 @@ import static lombok.AccessLevel.PROTECTED;
 public class ReviewDto {
 
     private String review;
-    private byte rating;
+    private float rating;
 }
